[WordSeg] Move support tests to SupportTests
Move support tests into Tests/SupportTests
Conform test names to swift-models convention
