+++ Tests/TextTests/WordSegmentationTests/WordSegmentationTests.swift
@@ -100,72 +100,3 @@ class WordSegSemiRingTests: XCTestCase {
   ]
 }
 
-class WordSegVocabularyTests: XCTestCase {
-  func test_AlphabetaConstruct() {
-    let characters: Alphabet = Alphabet(
-      [
-        "a",
-        "b",
-        "c",
-      ], eos: "</s>", eow: "</w>", pad: "<pad>")
-    // a, b, c, EOS, EOW, PAD
-    XCTAssertEqual(characters.count, 6)
-    XCTAssertEqual(characters.eos, 3)
-    XCTAssertEqual(characters.eow, 4)
-    XCTAssertEqual(characters.pad, 5)
-  }
-
-  func test_CharacterSequenceConstruct() {
-    let characters: Alphabet = Alphabet(
-      [
-        "a",
-        "c",
-        "t",
-      ], eos: "</s>", eow: "</w>", pad: "<pad>")
-    let cat: CharacterSequence? = try? CharacterSequence(
-      alphabet: characters, appendingEoSTo: "cat")
-    XCTAssertNotEqual(cat, nil)
-    // FIXME(abdulras) should the EoS be visible?
-    XCTAssertEqual(cat?.characters, [Int32(1), Int32(0), Int32(2), characters.eos])
-
-    let bat: CharacterSequence? = try? CharacterSequence(
-      alphabet: characters, appendingEoSTo: "bat")
-    XCTAssertEqual(bat, nil)
-  }
-
-  func test_LexiconConstruct() {
-    let characters: Alphabet = Alphabet(
-      [
-        "a", "b", "e", "g", "h", "l", "m", "p", "t",
-      ], eos: "</s>", eow: "</w>", pad: "<pad>")
-    let strings: Lexicon = Lexicon([
-      try! CharacterSequence(alphabet: characters, appendingEoSTo: "alpha"),
-      try! CharacterSequence(alphabet: characters, appendingEoSTo: "beta"),
-      try! CharacterSequence(alphabet: characters, appendingEoSTo: "gamma"),
-    ])
-
-    XCTAssertEqual(strings.count, 3)
-  }
-
-  func test_LexiconFromSequences() {
-    let alphabet: Alphabet = Alphabet(
-      [
-        "a", "b", "e", "g", "h", "l", "m", "p", "t",
-      ], eos: "</s>", eow: "</w>", pad: "<pad>")
-    let lexicon: Lexicon = Lexicon(
-      from: [
-        try! CharacterSequence(alphabet: alphabet, appendingEoSTo: "alpha"),
-        try! CharacterSequence(alphabet: alphabet, appendingEoSTo: "beta"),
-        try! CharacterSequence(alphabet: alphabet, appendingEoSTo: "gamma"),
-      ], alphabet: alphabet, maxLength: 5, minFreq: 4)
-
-    XCTAssertEqual(lexicon.count, 3)
-  }
-
-  static var allTests = [
-    ("test_AlphabetaConstruct", test_AlphabetaConstruct),
-    ("test_CharacterSequenceConstruct", test_CharacterSequenceConstruct),
-    ("test_LexiconConstruct", test_LexiconConstruct),
-    ("test_LexiconFromSequences", test_LexiconFromSequences),
-  ]
-}
