+++ Tests/TextTests/XCTestManifests.swift
@@ -20,7 +20,6 @@ import XCTest
       testCase(TextInferenceTests.allTests),
       testCase(WordSegDatasetTests.allTests),
       testCase(WordSegSemiRingTests.allTests),
-      testCase(WordSegVocabularyTests.allTests),
       testCase(WordSegProbeLayerTests.allTests),
     ]
   }
